+++ .gitattributes
@@ -1,0 +1,1 @@
+.git_archival.txt  export-subst
