+++ .gitignore
@@ -5,11 +5,12 @@
 /.idea/
 cspice
 dist/
-cspyce.egg-info/
+*cspyce.egg-info/
 **/*wrap.c
 wheelhouse/
 .DS_Store
 cspyce/cspyce0.py
 cspyce/cspyce0_info.py
+cspyce/_version.py
 swig/vectorize.i
 *venv*
