+++ pyproject.toml
@@ -1,0 +1,48 @@
+[build-system]
+requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "rfrenchseti_cspyce"
+dynamic = ["version"]
+description = "High-performance Python interface to the NAIF CSPICE library"
+readme = "README.md"
+requires-python = ">=3.8"
+dependencies = [
+    "numpy >= 1.17.0",
+]
+license = {text = "Apache-2.0"}
+authors = [
+  {name = "Mark R. Showalter", email = "mshowalter@seti.org"}
+]
+maintainers = [
+  {name = "Robert S. French", email = "rfrench@seti.org"}
+]
+keywords = ["cspyce", "spice", "cspice", "naif", "jpl", "space", "geometry",
+            "ephemeris"]
+classifiers = [
+  "Development Status :: 5 - Production/Stable",
+  "Natural Language :: English",
+  "Topic :: Scientific/Engineering",
+  "Topic :: Scientific/Engineering :: Astronomy",
+  "License :: OSI Approved :: Apache Software License",
+  "Programming Language :: Python :: 3.8",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Operating System :: MacOS :: MacOS X",
+  "Operating System :: POSIX :: Linux",
+  "Operating System :: Microsoft :: Windows"
+]
+
+[project.urls]
+homepage = "https://github.com/SETI/pds-cspyce"
+repository = "https://github.com/SETI/pds-cspyce"
+source = "https://github.com/SETI/pds-cspyce"
+issues = "https://github.com/SETI/pds-cspyce/issues"
+
+[tool.setuptools]
+packages = ["cspyce"]
+
+[tool.setuptools_scm]
+write_to = "cspyce/_version.py"
