--- setup.cfg
@@ -1,38 +1,0 @@
-[metadata]
-name = cspyce
-description = High-performance Python interface to the NAIF CSPICE library
-long_description = file: README.md
-long_description_content_type = text/markdown
-author = Mark R. Showalter
-author_email = mshowalter@seti.org
-maintainer = Robert S. French
-maintainer_email = rfrench@seti.org
-license = Apache-2.0
-url = https://github.com/SETI/pds-cspyce
-project_urls =
-    Homepage = https://github.com/SETI/pds-cspyce
-    Repository = https://github.com/SETI/pds-cspyce
-    Source = https://github.com/SETI/pds-cspyce
-    Bug Tracker = https://github.com/SETI/pds-cspyce/issues
-keywords = cspyce, spice, cspice, naif, jpl, space, geometry, ephemeris
-readme = README.md
-classifiers =
-     Development Status :: 5 - Production/Stable
-     Natural Language :: English
-     Topic :: Scientific/Engineering
-     Topic :: Scientific/Engineering :: Astronomy
-     License :: OSI Approved :: Apache Software License
-     Programming Language :: Python :: 3.8
-     Programming Language :: Python :: 3.9
-     Programming Language :: Python :: 3.10
-     Programming Language :: Python :: 3.11
-     Operating System :: MacOS :: MacOS X
-     Operating System :: POSIX :: Linux
-     Operating System :: Microsoft :: Windows
-
-[options]
-packages=cspyce
-python_requires =  >=3.6, <4
-install_requires =
-    numpy>=1.17.0
-tests_require = numpy>=1.17.0; pytest>=7.0.0; pandas>=0.24.0
